# openapi.yaml
openapi: 3.1.0
info:
  title: VoiceBlender API
  description: >
    VoiceBlender bridges SIP and WebRTC voice calls with multi-party audio
    mixing, real-time speech-to-text, text-to-speech, AI agent integration,
    recording, and webhook-based event delivery.


    **Configuration environment variables:**
    `INSTANCE_ID` (auto-generated UUID; included in all API response bodies and
    webhook events), `HTTP_ADDR` (:8080), `SIP_BIND_IP`, `SIP_PORT`, `SIP_HOST`,
    `ICE_SERVERS`, `RECORDING_DIR`, `LOG_LEVEL`, `WEBHOOK_URL`,
    `ELEVENLABS_API_KEY`, `VAPI_API_KEY` (VAPI agent provider API key),
    `S3_BUCKET`, `S3_REGION`, `S3_ENDPOINT`, `S3_PREFIX`.
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080/v1
    description: Local development server
...
paths:
...
  /legs:
    post:
      operationId: createLeg
      summary: Originate an outbound SIP call
      tags: [Legs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLegRequest'
      responses:
        '201':
          description: Leg created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Leg'
        '400':
          description: Invalid JSON, bad SIP URI, unknown codec, or unsupported type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      operationId: listLegs
      summary: List all active legs
      tags: [Legs]
      responses:
        '200':
          description: Array of legs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Leg'
  /legs/{id}:
    parameters:
      - $ref: '#/components/parameters/LegId'

    get:
      operationId: getLeg
      summary: Get a single leg
      tags: [Legs]
      responses:
        '200':
          description: Leg details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Leg'
        '404':
          $ref: '#/components/responses/LegNotFound'

    delete:
      operationId: deleteLeg
      summary: Hang up a leg
      tags: [Legs]
      responses:
        '200':
          description: Leg hung up
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: hung_up
        '404':
          $ref: '#/components/responses/LegNotFound'
  /legs/{id}/answer:
    parameters:
      - $ref: '#/components/parameters/LegId'
    post:
      operationId: answerLeg
      summary: Answer a ringing or early-media inbound SIP leg
      tags: [Legs]
      responses:
        '200':
          description: Answer initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: answering
        '400':
          description: Not a SIP inbound leg
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/LegNotFound'
        '409':
          description: Leg is not in ringing or early_media state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /legs/{id}/early-media:
    parameters:
      - $ref: '#/components/parameters/LegId'
    post:
      operationId: earlyMediaLeg
      summary: Enable early media on a ringing inbound SIP leg
      description: >
        Sends SIP 183 Session Progress with SDP and sets up the media pipeline.
        The leg transitions to `early_media` state, allowing audio playback and
        room participation before the call is answered.
      tags: [Legs]
      responses:
        '200':
          description: Early media enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: early_media
        '400':
          description: Not a SIP inbound leg
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/LegNotFound'
        '409':
          description: Leg is not in ringing state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Media setup failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /legs/{id}/mute:
    parameters:
      - $ref: '#/components/parameters/LegId'
    post:
      operationId: muteLeg
      summary: Mute a leg
      description: >
        A muted leg's audio is excluded from the room mix and speaking events
        are suppressed. Taps (recording/STT) still receive the muted leg's own audio.
      tags: [Legs]
      responses:
        '200':
          description: Leg muted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: muted
        '404':
          $ref: '#/components/responses/LegNotFound'

    delete:
      operationId: unmuteLeg
      summary: Unmute a leg
      tags: [Legs]
      responses:
        '200':
          description: Leg unmuted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: unmuted
        '404':
          $ref: '#/components/responses/LegNotFound'
  /legs/{id}/hold:
    parameters:
      - $ref: '#/components/parameters/LegId'
    post:
      operationId: holdLeg
      summary: Put a SIP call on hold
      description: >
        Sends a re-INVITE with `sendonly` SDP direction. The RTP timeout is
        paused while held, and a 2-hour auto-hangup timer starts.
      tags: [Legs]
      responses:
        '200':
          description: Call held
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: held
        '400':
          description: Not a SIP leg
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/LegNotFound'
        '409':
          description: Leg is not in connected state, or already held
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: unholdLeg
      summary: Resume a held SIP call
      description: Sends a re-INVITE with `sendrecv` SDP direction.
      tags: [Legs]
      responses:
        '200':
          description: Call resumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: resumed
        '400':
          description: Not a SIP leg
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/LegNotFound'
        '409':
          description: Leg is not held
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /legs/{id}/dtmf:
    parameters:
      - $ref: '#/components/parameters/LegId'
    post:
      operationId: sendDTMF
      summary: Send DTMF digits on a leg
      tags: [Legs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DTMFRequest'
      responses:
        '200':
          description: Digits sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: sent
        '400':
          description: Invalid JSON or empty digits
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/LegNotFound'
        '500':
          description: DTMF writer unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /legs/{id}/play:
    parameters:
      - $ref: '#/components/parameters/LegId'
    post:
      operationId: playLeg
      summary: Start audio playback to a leg
      tags: [Legs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaybackRequest'
      responses:
        '200':
          description: Playback started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybackResponse'
        '400':
          description: Invalid JSON or volume out of range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/LegNotFound'
        '409':
          description: Leg has no audio writer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /legs/{id}/play/{playbackID}:
    parameters:
      - $ref: '#/components/parameters/LegId'
      - $ref: '#/components/parameters/PlaybackId'
    patch:
      operationId: volumePlayLeg
      summary: Change the volume of an active leg playback
      tags: [Legs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VolumeRequest'
      responses:
        '200':
          description: Volume updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: ok
        '400':
          description: Invalid JSON or volume out of range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Playback not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      operationId: stopPlayLeg
      summary: Stop audio playback on a leg
      tags: [Legs]
      responses:
        '200':
          description: Playback stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: stopped
        '404':
          description: No playback in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /legs/{id}/tts:
    parameters:
      - $ref: '#/components/parameters/LegId'
    post:
      operationId: ttsLeg
      summary: Synthesize speech and play it on a leg
      tags: [Legs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TTSRequest'
      responses:
        '200':
          description: TTS playback started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TTSResponse'
        '400':
          description: Invalid JSON, missing text/voice, or volume out of range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/LegNotFound'
        '409':
          description: Leg has no audio writer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: No API key provided for the selected provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /legs/{id}/record:
    parameters:
      - $ref: '#/components/parameters/LegId'
    post:
      operationId: recordLeg
      summary: Start recording a leg to a WAV file
      description: >
        For SIP legs, recording is stereo (left=incoming, right=outgoing).
        For legs in a room, stereo at 16kHz (left=participant audio, right=mixed-minus-self).
      tags: [Legs]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordingRequest'
      responses:
        '200':
          description: Recording started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingResponse'
        '400':
          description: Invalid storage type, S3 not configured, or invalid S3 credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/LegNotFound'
        '409':
          description: Leg has no audio reader or room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to create recording file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: stopRecordLeg
      summary: Stop recording a leg
      tags: [Legs]
      responses:
        '200':
          description: Recording stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingStopResponse'
        '404':
          description: No recording in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /legs/{id}/stt:
    parameters:
      - $ref: '#/components/parameters/LegId'
    post:
      operationId: sttLeg
      summary: Start real-time speech-to-text on a leg
      tags: [Legs]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/STTRequest'
      responses:
        '200':
          description: STT started
          content:
            application/json:
              schema:
                type: object
                properties:
                  instance_id:
                    type: string
                    description: Instance identifier
                  status:
                    type: string
                    example: stt_started
                  leg_id:
                    type: string
                required: [status, leg_id]
        '404':
          $ref: '#/components/responses/LegNotFound'
        '409':
          description: Leg not connected, STT already running, or no audio reader
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: No ElevenLabs API key provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: stopSTTLeg
      summary: Stop speech-to-text on a leg
      tags: [Legs]
      responses:
        '200':
          description: STT stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: stt_stopped
        '404':
          description: No STT in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /legs/{id}/agent:
    parameters:
      - $ref: '#/components/parameters/LegId'
    post:
      operationId: agentLeg
      summary: Attach an AI agent to a leg
      description: >
        Bridges audio bidirectionally with a conversational AI agent.
        Supported providers: `elevenlabs` (default), `vapi`, `pipecat`.
        Standalone legs use direct audio; legs in a room use mixer taps.
      tags: [Legs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRequest'
      responses:
        '200':
          description: Agent started
          content:
            application/json:
              schema:
                type: object
                properties:
                  instance_id:
                    type: string
                    description: Instance identifier
                  status:
                    type: string
                    example: agent_started
                  leg_id:
                    type: string
                required: [status, leg_id]
        '400':
          description: Invalid JSON or missing agent_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/LegNotFound'
        '409':
          description: Leg not connected, agent already attached, or no audio reader/writer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: No API key provided for the selected provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: stopAgentLeg
      summary: Detach the agent from a leg
      tags: [Legs]
      responses:
        '200':
          description: Agent stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: agent_stopped
        '404':
          description: No agent attached to this leg
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Rooms ───────────────────────────────────────────────────────────────
  /rooms:
    post:
      operationId: createRoom
      summary: Create a room
      tags: [Rooms]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomCreateRequest'
      responses:
        '201':
          description: Room created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '409':
          description: Room ID already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      operationId: listRooms
      summary: List all rooms with participants
      tags: [Rooms]
      responses:
        '200':
          description: Array of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
  /rooms/{id}:
    parameters:
      - $ref: '#/components/parameters/RoomId'

    get:
      operationId: getRoom
      summary: Get a room with participants
      tags: [Rooms]
      responses:
        '200':
          description: Room details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '404':
          $ref: '#/components/responses/RoomNotFound'

    delete:
      operationId: deleteRoom
      summary: Delete a room
      tags: [Rooms]
      responses:
        '200':
          description: Room deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: deleted
        '404':
          $ref: '#/components/responses/RoomNotFound'
  /rooms/{id}/legs:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    post:
      operationId: addLegToRoom
      summary: Add or move a leg to a room
      description: >
        Add a leg to a room (auto-creates room if it doesn't exist).
        If the leg is already in a different room, it is atomically moved
        to the target room. A ringing inbound SIP leg is automatically
        answered before being added.
      tags: [Rooms]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddLegRequest'
      responses:
        '200':
          description: Leg added or moved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: added
        '400':
          description: Invalid JSON, leg not found, or leg not connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /rooms/{id}/legs/{legID}:
    parameters:
      - $ref: '#/components/parameters/RoomId'
      - name: legID
        in: path
        required: true
        schema:
          type: string
        description: Leg ID
    put:
      operationId: moveLegToRoom
      summary: Move a leg to a different room
      tags: [Rooms]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                to_room_id:
                  type: string
                  description: Destination room ID
              required: [to_room_id]
      responses:
        '200':
          description: Leg moved
          content:
            application/json:
              schema:
                type: object
                properties:
                  instance_id:
                    type: string
                    description: Instance identifier
                  status:
                    type: string
                    example: moved
                  from:
                    type: string
                    description: Source room ID
                  to:
                    type: string
                    description: Destination room ID
                required: [status, from, to]
        '400':
          description: Invalid JSON, leg not found, or move not possible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: removeLegFromRoom
      summary: Remove a leg from a room
      tags: [Rooms]
      responses:
        '200':
          description: Leg removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: removed
        '400':
          description: Room or leg not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /rooms/{id}/play:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    post:
      operationId: playRoom
      summary: Play audio to a room
      tags: [Rooms]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaybackRequest'
      responses:
        '200':
          description: Playback started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybackResponse'
        '400':
          description: Invalid JSON or volume out of range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/RoomNotFound'
        '409':
          description: Room has no participants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /rooms/{id}/play/{playbackID}:
    parameters:
      - $ref: '#/components/parameters/RoomId'
      - $ref: '#/components/parameters/PlaybackId'
    patch:
      operationId: volumePlayRoom
      summary: Change the volume of an active room playback
      tags: [Rooms]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VolumeRequest'
      responses:
        '200':
          description: Volume updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: ok
        '400':
          description: Invalid JSON or volume out of range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Playback not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      operationId: stopPlayRoom
      summary: Stop room playback
      tags: [Rooms]
      responses:
        '200':
          description: Playback stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: stopped
        '404':
          description: No playback in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /rooms/{id}/tts:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    post:
      operationId: ttsRoom
      summary: Synthesize speech and play it into a room
      tags: [Rooms]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TTSRequest'
      responses:
        '200':
          description: TTS playback started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TTSResponse'
        '400':
          description: Invalid JSON, missing text/voice, or volume out of range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/RoomNotFound'
        '409':
          description: Room has no participants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: No API key provided for the selected provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /rooms/{id}/ws:
    get:
      operationId: wsRoom
      summary: WebSocket audio stream for a room
      tags: [Rooms]
      description: >
        Upgrades to a WebSocket connection and joins the room as a bidirectional
        audio participant. The client sends and receives 16kHz 16-bit signed
        little-endian PCM audio (mono), base64-encoded in JSON text frames.
        Each audio frame is 640 bytes (20ms).
      parameters:
        - $ref: '#/components/parameters/RoomId'
      responses:
        '101':
          description: >
            WebSocket upgrade successful. Server sends a `connected` message
            followed by mixed-minus-self audio frames.
        '404':
          description: Room not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── WebRTC ──────────────────────────────────────────────────────────────
...
  /webhooks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Webhook ID
    delete:
      operationId: deleteWebhook
      summary: Unregister a webhook
      tags: [Webhooks]
      responses:
        '200':
          description: Webhook deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: deleted
        '404':
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Observability ───────────────────────────────────────────────────────────
...
components:
...
  schemas:
...
    Leg:
      type: object
      properties:
        instance_id:
          type: string
          description: Instance identifier
        leg_id:
          type: string
          description: Unique leg identifier (UUID)
        type:
          type: string
          enum: [sip_inbound, sip_outbound, webrtc]
        state:
          type: string
          enum: [ringing, early_media, connected, held, hung_up]
        room_id:
          type: string
          description: Room ID if the leg is in a room, empty otherwise
        muted:
          type: boolean
          description: Whether the leg is muted
        held:
          type: boolean
          description: Whether the call is on hold (SIP legs only)
        sip_headers:
          type: object
          additionalProperties:
            type: string
          description: 'X-* headers from the inbound INVITE. Only present on sip_inbound legs.'
      required: [leg_id, type, state]
...
    Error:
      type: object
      properties:
        instance_id:
          type: string
          description: Instance identifier
        error:
          type: string
          description: Error message
      required: [error]
...